fix: page count data를 long으로 변경

+++ be/src/main/java/kr/codesquad/jazzmeet/venue/dto/response/VenueSearchResponse.java
@@ -8,9 +8,9 @@ import lombok.Builder;
 @Builder
 public record VenueSearchResponse(
 	List<VenueSearch> venues,
-	int venueCount,
+	long venueCount,
 	int currentPage,
-	int maxPage) {
+	long maxPage) {
 	public static VenueSearchResponse emptyVenues() {
 		return VenueSearchResponse.builder().venues(List.of()).venueCount(0).currentPage(1).maxPage(1).build();
 	}
